requirements.txt
streamlit==1.31.0
pandas==2.1.4
plotly==5.18.0
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0